All main tests work
