+++ tests/conftest.py
@@ -1,3 +1,5 @@
 import os
 
 os.environ["AVDB_USE_LOCKING"] = "1"
+
+pytest_plugins = ("pytest_asyncio",)
